refactor: Se ocupa hook
En el aparatado de ThemeSwitch y Form
